add img,confirmDiag,guiTemp

+++ footTunisie/src/edu/db3a4/gui2/AffichageTournoiController.java
@@ -11,6 +11,7 @@ import edu.db3a4.tools.MyConnection;
 import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
@@ -30,7 +31,9 @@ import javafx.collections.ObservableList;
 import javafx.embed.swing.SwingFXUtils;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
+import javafx.scene.Parent;
 import javafx.scene.control.Button;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.DatePicker;
@@ -40,6 +43,8 @@ import javafx.scene.control.TextField;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
+import javafx.scene.input.KeyEvent;
+import javafx.scene.input.MouseEvent;
 import javax.imageio.ImageIO;
 import javax.swing.ImageIcon;
 import javax.swing.JOptionPane;
@@ -49,13 +54,6 @@ import javax.swing.JOptionPane;
  *
  * @author ASUS
  */
-
-
-
-
-
-
-
 
 public class AffichageTournoiController implements Initializable {
 
@@ -90,9 +88,16 @@ public class AffichageTournoiController implements Initializable {
     @FXML
     private DatePicker pickerDate;
     @FXML
-    private Button ajout;
-    @FXML
     private ImageView image;
+    @FXML
+    private Button gestion;
+    @FXML
+    private ImageView logoImg;
+    @FXML
+    private Button btnAjout1;
+    @FXML
+    private TableColumn<Tournoi, String> tbImg;
+    @FXML
 
     /**
      * Initializes the controller class.
@@ -102,7 +107,15 @@ public class AffichageTournoiController implements Initializable {
     @Override
     public void initialize(URL url, ResourceBundle rb) {
                 tf_nbrE.getItems().addAll(4,6,8,10);
+                
+               Image img;
+  try {
+            img = new Image(new FileInputStream("C:\\Users\\ASUS\\Desktop\\git\\Pidev\\footTunisie\\src\\images\\logo.png"));
+              logoImg.setImage(img);
 
+        } catch (FileNotFoundException ex) {
+            Logger.getLogger(AddTournoiController.class.getName()).log(Level.SEVERE, null, ex);
+        }
        try {
             String requete = "SELECT nom FROM  terrain";
             Statement st = MyConnection.getInstance().getCnx()
@@ -121,8 +134,9 @@ public class AffichageTournoiController implements Initializable {
             nbE.setCellValueFactory(new PropertyValueFactory<>("nbr_equipe"));
             DTournoi.setCellValueFactory(new PropertyValueFactory<>("dateTournoi"));
             Ttournoi.setCellValueFactory(new PropertyValueFactory<>("terrainTournoi"));
-                                                                                                                                              
+            tbImg.setCellValueFactory(new PropertyValueFactory<>("image"));                                                                                                                             
             afficher.setItems(pcd.displayTournoi());
+           
                  
     }
 
@@ -144,17 +158,20 @@ public class AffichageTournoiController implements Initializable {
 
     @FXML
     private void supprimerTournoi(ActionEvent event) {
-         String id = tf_id.getText();
+        int opt = JOptionPane.showConfirmDialog(null, "Confirmer la suppression ?","Supprimer",JOptionPane.YES_NO_OPTION);
+      if(opt==0){
+        String id = tf_id.getText();
          Integer id1 = Integer.parseInt(id);
             TournoiCRUD pcd = new TournoiCRUD();
             pcd.supprimerTournoi(id1);
-            afficher.setItems(pcd.displayTournoi());
-            
+            afficher.setItems(pcd.displayTournoi());}
     }
 
     @FXML
     private void modifierTournoi(ActionEvent event) {
-         String id = tf_id.getText();
+         int opt = JOptionPane.showConfirmDialog(null, "Confirmer la modification ?","Modifier",JOptionPane.YES_NO_OPTION);
+      if(opt==0){
+        String id = tf_id.getText();
          Integer id1 = Integer.parseInt(id);
          String nomT = tf_nom.getText();
             String nomTe = cmbTerrain.getValue();
@@ -162,8 +179,33 @@ public class AffichageTournoiController implements Initializable {
             LocalDate dateT = pickerDate.getValue();
             TournoiCRUD pcd = new TournoiCRUD();
             pcd.updateTournoi(id1,nomT,nbr_Equipe,dateT,nomTe);
-            afficher.setItems(pcd.displayTournoi());
-        
+            afficher.setItems(pcd.displayTournoi());   }
+                            
     }
+
+    @FXML
+    private void gestion(ActionEvent event) {
+    }
+
+    @FXML
+    private void ajouter(ActionEvent event) throws IOException {
+        FXMLLoader loader = new FXMLLoader(getClass()
+                    .getResource("AddTournoi.fxml"));
+            Parent root = loader.load();
+            tf_id.getScene().setRoot(root);
+    }
+    @FXML
+    private void tabV(MouseEvent event) throws FileNotFoundException {
+         Tournoi tournoi = afficher.getSelectionModel().getSelectedItem();
+            tf_id.setText(String.valueOf(tournoi.getId()));
+            tf_nom.setText(tournoi.getNomTournoi());
+            tf_nbrE.setValue(tournoi.getNbr_equipe());
+            pickerDate.setValue(tournoi.getDateTournoi());
+            cmbTerrain.setValue(tournoi.getTerrainTournoi());
+            Image img = new Image(new FileInputStream("C:\\Users\\ASUS\\Desktop\\git\\Pidev\\footTunisie\\src\\images\\"+tournoi.getImage()));
+            image.setImage(img);
+    }
+
+  
     }
